+++ CHANGES.rst
@@ -414,6 +414,11 @@ API Changes
   - The repr() function of ``astropy.table.Table`` now shows the units
     if any columns have the unit attribute set.  [#2180]
 
+  - The semantics of the config options ``table.max_lines`` and
+    ``table.max_width`` has changed slightly.  If these values are not
+    set in the config file, astropy will try to determine the size
+    automatically from the terminal. [#2683]
+
 - ``astropy.time``
 
   - Correct use of UT in TDB calculation [#1938, #1939].